# gen/openapi/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.1 DO NOT EDIT.
package gen

// BodyLoginLoginPost defines model for Body_login_login_post.
type BodyLoginLoginPost struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// HTTPValidationError defines model for HTTPValidationError.
type HTTPValidationError struct {
	Detail *[]ValidationError `json:"detail,omitempty"`
}

// MathQuery defines model for MathQuery.
type MathQuery struct {
	A int `json:"a"`
	B int `json:"b"`
}

// MathResponse defines model for MathResponse.
type MathResponse struct {
	Query  *MathQuery `json:"query,omitempty"`
	Result *int       `json:"result,omitempty"`
}

// MessageResponse defines model for MessageResponse.
type MessageResponse struct {
	Code *int    `json:"code,omitempty"`
	Msg  *string `json:"msg,omitempty"`
}

// ValidationError defines model for ValidationError.
type ValidationError struct {
	Loc  []string `json:"loc"`
	Msg  string   `json:"msg"`
	Type string   `json:"type"`
}

// ValueResponse defines model for ValueResponse.
type ValueResponse struct {
	Code *int `json:"code,omitempty"`
	Msg  *int `json:"msg,omitempty"`
}

// ReadModuloModuloPostJSONBody defines parameters for ReadModuloModuloPost.
type ReadModuloModuloPostJSONBody MathQuery

// ReadModuloModuloPostJSONRequestBody defines body for ReadModuloModuloPost for application/json ContentType.
type ReadModuloModuloPostJSONRequestBody ReadModuloModuloPostJSONBody